+++ src/main/java/com/jayway/controller/Transfer.java
@@ -6,41 +6,41 @@ import javax.validation.constraints.NotNull;
 class Transfer {
 
     @NotNull(message = "From account number must be set")
-    private Long fromAccountNumber;
+    private Integer fromAccountNumber;
 
     @NotNull(message = "To account number must be set")
-    private Long toAccountNumber;
+    private Integer toAccountNumber;
 
     @Min(message = "Amount must be >= 0", value = 0)
-    private long amount;
+    private int amount;
 
 
-    Long getFromAccountNumber() {
+    Integer getFromAccountNumber() {
         return fromAccountNumber;
     }
 
 
-    void setFromAccountNumber(Long fromAccountNumber) {
+    void setFromAccountNumber(Integer fromAccountNumber) {
         this.fromAccountNumber = fromAccountNumber;
     }
 
 
-    Long getToAccountNumber() {
+    Integer getToAccountNumber() {
         return toAccountNumber;
     }
 
 
-    void setToAccountNumber(Long toAccountNumber) {
+    void setToAccountNumber(Integer toAccountNumber) {
         this.toAccountNumber = toAccountNumber;
     }
 
 
-    long getAmount() {
+    int getAmount() {
         return amount;
     }
 
 
-    void setAmount(long amount) {
+    void setAmount(int amount) {
         this.amount = amount;
     }
 }
