+++ src/main/java/com/jayway/domain/Account.java
@@ -3,7 +3,7 @@ package com.jayway.domain;
 
 public interface Account {
 
-    Long getAccountNumber();
+    Integer getAccountNumber();
 
-    long getBalance();
+    int getBalance();
 }
