+++ src/main/java/com/jayway/service/UnknownAccountException.java
@@ -8,11 +8,11 @@ public class UnknownAccountException extends RuntimeException {
 
     private static final long serialVersionUID = -5562010974917511663L;
 
-    public UnknownAccountException(long accountNumber) {
+    public UnknownAccountException(Integer accountNumber) {
         super("Unknown account number: " + accountNumber);
     }
 
-    public UnknownAccountException(long accountNumber, Exception cause) {
+    public UnknownAccountException(Integer accountNumber, Exception cause) {
         super("Unknown account number: " + accountNumber, cause);
     }
 }
